fix(*): merge all the profiles into one and only
That was a test I've conducted for quite long time, but I've not much
time to have this large massive configuration management and neither
have time to rice multiple machine. Reality is that I only use the
daedalus device and the others I maybe use 10~20min every months, so not
worth spent times in configuration management anymore.